# docs/contracts/1.6/06_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Smart Wardrobe Prototype API
  version: 1.6.0
  description: HLD v1.1 / FSD v1.1 / Backend LLD v1.1 기반. 서버 소유의 REST 계약; Decart·ThinQ
    CLO 외부 실제 API 규격을 주장하지 않음. Bearer 인증과 demo_mode 세션은 프로토타입 정책이며 운영 보안 정책이 아님.
servers:
- url: http://localhost:8000
  description: Local FastAPI prototype
...
security:
- BearerAuth: []
paths:
  /api/v1/sessions:
    post:
      tags:
      - Common
      summary: 프로필 선택 및 세션 시작
      operationId: post_api_v1_sessions
      x-fsd-ids:
      - COM-001
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionStartRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStartResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security: []
...
  /api/v1/garments:
    get:
      tags:
      - Garments
      summary: 의류 목록 필터링
      operationId: get_api_v1_garments
      x-fsd-ids:
      - GAR-001
      parameters:
      - *id001
      - name: owner_id
        in: query
        required: false
        schema: *id002
      - name: category
        in: query
        required: false
        schema: &id003
          type: string
      - name: color
        in: query
        required: false
        schema: *id003
      - name: season
        in: query
        required: false
        schema: *id003
      - name: status
        in: query
        required: false
        schema: *id003
      - &id004
        name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      - &id005
        name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GarmentList'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags:
      - Garments
      summary: 의류 등록
      operationId: post_api_v1_garments
      x-fsd-ids:
      - GAR-003
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GarmentUpsert'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/garments/{garment_id}:
    delete:
      tags:
      - Garments
      summary: 의류 삭제 (버전 검증, soft delete)
      operationId: delete_api_v1_garments_garment_id
      x-fsd-ids:
      - GAR-003
      parameters:
      - name: garment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
        description: 버전 충돌 방지
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GarmentDeleted'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags:
      - Garments
      summary: 의류 상세·위치
      operationId: get_api_v1_garments_garment_id
      x-fsd-ids:
      - GAR-002
      parameters:
      - name: garment_id
        in: path
        required: true
        schema: *id002
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags:
      - Garments
      summary: 의류 정보 변경
      operationId: patch_api_v1_garments_garment_id
      x-fsd-ids:
      - GAR-003
      parameters:
      - name: garment_id
        in: path
        required: true
        schema: *id002
      - name: If-Match
        in: header
        required: true
        schema: *id003
        description: 버전 충돌 방지
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GarmentUpsert'
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Garment'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/garment-observations:
    post:
      tags:
      - Recognition
      summary: 의류 관측 등록 및 상태 융합
      operationId: post_api_v1_garment_observations
      x-fsd-ids:
      - GAR-004
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/garments/locate:
    post:
      tags:
      - Storage
      summary: 구조화 조건 기반 의류 위치 조회
      operationId: post_api_v1_garments_locate
      x-fsd-ids:
      - STO-001
      description: ThinQ CLO 또는 Mock이 구조화한 조건만 처리. 소유자 권한은 서버에서 검증.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GarmentLocateRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocatedGarmentList'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/settings:
    get:
      tags:
      - Settings
      summary: 사용자·연동 설정 조회
      operationId: get_api_v1_settings
      x-fsd-ids:
      - SET-001
      parameters:
      - *id001
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags:
      - Settings
      summary: 사용자·연동 설정 변경
      operationId: put_api_v1_settings
      x-fsd-ids:
      - SET-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/assets/upload-intents:
    post:
      tags:
      - Assets
      summary: 직접 업로드용 임시 URL 발급
      operationId: post_api_v1_assets_upload_intents
      x-fsd-ids:
      - GAR-003
      - VTON-002
      - STYLE-001
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadIntentRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadIntentResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/assets/{asset_id}/finalize:
    post:
      tags:
      - Assets
      summary: 업로드 완료 및 무결성 확인
      operationId: post_api_v1_assets_asset_id_finalize
      x-fsd-ids:
      - GAR-003
      - VTON-002
      - STYLE-001
      parameters:
      - name: asset_id
        in: path
        required: true
        schema: *id002
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetFinalizeRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/settings/consent-history:
    get:
      tags:
      - Settings
      summary: 본인 이미지 동의 변경 이력
      operationId: get_api_v1_settings_consent_history
      x-fsd-ids:
      - SET-001
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentHistory'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 서버가 발급한 사용자 범위 세션 토큰. 요청 member_id/household_id를 신뢰하지 않고 인증 주체와
        비교.
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/Error'
      required:
      - error
      additionalProperties: false
...
    SessionStartRequest:
      type: object
      properties:
        household_id: *id002
        member_id: *id002
        demo_mode:
          type: boolean
      required:
      - household_id
      - member_id
      additionalProperties: false
    SessionStartResponse:
      type: object
      properties:
        session_id: *id002
        access_token: *id003
        token_type:
          type: string
          enum:
          - Bearer
        expires_at: &id007
          type: string
          format: date-time
        member:
          $ref: '#/components/schemas/Member'
        capabilities:
          type: array
          items: *id003
      required:
      - session_id
      - access_token
      - token_type
      - expires_at
      - member
      - capabilities
      additionalProperties: false
...
    ObservationResponse:
      type: object
      properties:
        observation_id: *id002
        state_updated:
          type: boolean
        garment:
          $ref: '#/components/schemas/Garment'
      required:
      - observation_id
      - state_updated
      - garment
      additionalProperties: false
...
    GarmentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Garment'
        page:
          $ref: '#/components/schemas/Page'
      required:
      - items
      - page
      additionalProperties: false
...
    GarmentDeleted:
      additionalProperties: false
      properties:
        garment_id:
          format: uuid
          type: string
        deleted:
          const: true
          type: boolean
      required:
      - garment_id
      - deleted
      type: object
...